move differential tests into their own files
